+++ amethyst_network/src/simulation/events.rs
@@ -1,5 +1,6 @@
+use crate::simulation::Message;
 use bytes::Bytes;
-use std::net::SocketAddr;
+use std::{io, net::SocketAddr};
 
 /// Events which can be received from the network.
 #[derive(Debug)]
@@ -10,4 +11,10 @@ pub enum NetworkSimulationEvent {
     Connect(SocketAddr),
     // A host has disconnected from us
     Disconnect(SocketAddr),
+    // An error occurred while receiving a message.
+    RecvError(io::Error),
+    // An error occurred while sending a message.
+    SendError(io::Error, Message),
+    // An error occurred while managing connections.
+    ConnectionError(io::Error, Option<SocketAddr>),
 }