+++ amethyst_network/src/simulation/transport/tcp.rs
@@ -2,13 +2,13 @@
 
 use crate::simulation::{
     events::NetworkSimulationEvent,
+    message::Message,
     requirements::DeliveryRequirement,
     timing::{NetworkSimulationTime, NetworkSimulationTimeSystem},
     transport::{
         TransportResource, NETWORK_RECV_SYSTEM_NAME, NETWORK_SEND_SYSTEM_NAME,
         NETWORK_SIM_TIME_SYSTEM_NAME,
     },
-    Message,
 };
 use amethyst_core::{
     bundle::SystemBundle,
@@ -17,7 +17,7 @@ use amethyst_core::{
 };
 use amethyst_error::Error;
 use bytes::Bytes;
-use log::{error, warn};
+use log::warn;
 use std::{
     collections::HashMap,
     io::{self, Read as IORead, Write as IOWrite},
@@ -97,10 +97,10 @@ impl<'s> System<'s> for TcpStreamManagementSystem {
                 let s = match TcpStream::connect(message.destination) {
                     Ok(s) => s,
                     Err(e) => {
-                        warn!(
-                            "Error attempting to connection to {}: {:?}",
-                            message.destination, e
-                        );
+                        event_channel.single_write(NetworkSimulationEvent::ConnectionError(
+                            e,
+                            Some(message.destination),
+                        ));
                         return;
                     }
                 };
@@ -146,7 +146,8 @@ impl<'s> System<'s> for TcpConnectionListenerSystem {
                         break;
                     }
                     Err(e) => {
-                        error!("Error listening for connections: {}", e);
+                        event_channel
+                            .single_write(NetworkSimulationEvent::ConnectionError(e, None));
                         break;
                     }
                 };
@@ -163,18 +164,19 @@ impl<'s> System<'s> for TcpNetworkSendSystem {
         Write<'s, TransportResource>,
         Write<'s, TcpNetworkResource>,
         Read<'s, NetworkSimulationTime>,
+        Write<'s, EventChannel<NetworkSimulationEvent>>,
     );
 
-    fn run(&mut self, (mut transport, mut net, sim_time): Self::SystemData) {
+    fn run(&mut self, (mut transport, mut net, sim_time, mut channel): Self::SystemData) {
         let messages = transport.drain_messages_to_send(|_| sim_time.should_send_message_now());
-        for message in messages.iter() {
+        for message in messages {
             match message.delivery {
                 DeliveryRequirement::ReliableOrdered(Some(_)) => {
                     warn!("Streams are not supported by TCP and will be ignored.");
-                    write_message(message, &mut net);
+                    write_message(message, &mut net, &mut channel);
                 }
                 DeliveryRequirement::ReliableOrdered(_) | DeliveryRequirement::Default => {
-                    write_message(message, &mut net);
+                    write_message(message, &mut net, &mut channel);
                 }
                 delivery => panic!(
                     "{:?} is unsupported. TCP only supports ReliableOrdered by design.",
@@ -185,13 +187,14 @@ impl<'s> System<'s> for TcpNetworkSendSystem {
     }
 }
 
-fn write_message(message: &Message, net: &mut TcpNetworkResource) {
+fn write_message(
+    message: Message,
+    net: &mut TcpNetworkResource,
+    channel: &mut EventChannel<NetworkSimulationEvent>,
+) {
     if let Some((_, stream)) = net.get_stream(message.destination) {
         if let Err(e) = stream.write(&message.payload) {
-            error!(
-                "There was an error when attempting to send message: {:?}",
-                e
-            );
+            channel.single_write(NetworkSimulationEvent::SendError(e, message));
         }
     }
 }
@@ -239,7 +242,9 @@ impl<'s> System<'s> for TcpNetworkRecvSystem {
                                 *active = false;
                             }
                             io::ErrorKind::WouldBlock => {}
-                            _ => error!("Encountered an error receiving data: {:?}", e),
+                            _ => {
+                                event_channel.single_write(NetworkSimulationEvent::RecvError(e));
+                            }
                         }
                         break;
                     }